+++ R/rTwig-package.R
@@ -5,7 +5,9 @@
 #' @import rgl
 #' @import rlang
 #' @import tidytable
+#' @import Rdpack
 #' @import Rcpp
+#' @import RcppArmadillo
 #' @importFrom cobs cobs
 #' @importFrom colourvalues color_values
 #' @importFrom geometry convhulln